requirements.txt
flask==2.0.1
flask-cors==3.0.10
torch==2.2.0
torchvision==0.17.0
Pillow==10.2.0
transformers==4.37.2
numpy
tqdm==4.66.2
scikit-learn
requests==2.31.0
opencv-python==4.9.0.80
pandas==2.2.0
matplotlib==3.8.3
nltk==3.8.1
